pyproject.toml
[project]
name = "ai-podcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litserve>=0.2.6",
    "openai>=1.63.0",
    "playwright>=1.50.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "termcolor>=2.5.0",
]
[tool.setuptools]
py-modules = []
